Concrete/new/mnist/original_net/requirements.txt
concrete-ml==1.7
brevitas
torchvision
torch
onnx
onnxoptimizer
